refactor(playground): extract usePlaygroundFeedColumnLoader
